Move platform-specific code into its own file
Rather than having the JRuby-specific code in Argon2id::Password, move
it to a dedicated extension file that is responsible for either loading
the C extension or defining the JRuby version of the private API.

Take a lead from Nokogiri and move the compiled extension into
lib/argon2id rather than lib/.
